separate config code into its own package (#375)
