+++ config/config.go
@@ -1,4 +1,4 @@
-package types
+package config
 
 import (
 	"encoding/json"
